authors_app/serializers.py
from .models import Authors, Articles
from rest_framework import serializers


class AuthorsSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = Authors
        fields = '__all__'


class ArticlesSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = Articles
        fields = '__all__'
