requirements.txt
Django~=2.0.6
psycopg2-binary
django-extensions
djangorestframework
